ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.8.3
